// src/app/model/property/add.property.model.ts
export class AddPropertyModel {
    propertyType: string;
    adType: string;
    numberOfRoom: string;
    numberOfBathroom: string;
    numberOfReception: string;
    floorNumber: string;
    streetWidth: string;
    area: string;
    finishType: string;
    price: string;
    details: string;
    adTitle: string;
    hasGarage: string;
    latitude: string;
    city: string; 
    longitude: string;
    region: string;
    address: string;
    receptions: string;
}